Description of the Sun Moon Signal and how to calculate it
The sun moon signal can is simply the gravitational acceleration from the sun acting on the instrument minus the sun acting on the center of the earth.  This is added to the moon acting on the station minus the moon acting on the center of the earth.  This simple approximation, give a good fit to the signal measured by the superconducting gravimeter instruments, the broadband seismometers, and many many different kinds of gravimeter and instruments which have reached a critical sensitivity, so they they must account for the chaning gravitational potential, and its associated gradient at the instrument position.